moving RoutableHandlerResolver to a sub namespace
